Add guards to prevent exceptions while parsing JSON payload
OpenWeatherMap reports as much information as possible at the
time of the request, but some expected information might be
missing. This patch adds guards to prevent parsing exceptions
when some of the fields are missing in the JSON payload.

Additional this patch updates the disclaimer string to
properly state the purpose of this app

Change-Id: I5c18037d2de8f7ce07ed37e319b03584dcf922ae
TICKET: CYNGNOS-2606